refactor(review): TabList.vue を切り出して APG tabs ロジックをカプセル化
- ARIA 属性（role/aria-selected/aria-controls/aria-labelledby）と
  useId() による id 発行をすべて TabList 内に隠蔽
- Roving tabindex とキーボード（←/→/Home/End、自動アクティベーション）
  も内部で完結
- 親（ReviewProgression）は tabs / activeId / change だけを扱う

generic 型 `<T extends { id: string; label: string; sub? }>` で
任意のタブモデルに対応可能。tabClass コールバックで装飾の差し込み。
tabs.length<=1 ならタブリストを省略してパネルだけ描画する。
